refactor: complete DecimalData migration from Int to Long
Change price and cost fields in entities and UI from Int to Long
for consistency with the DecimalData refactor.

+++ database/src/desktopMain/kotlin/ru/pavlig43/database/data/batch/BatchCostPriceEntity.kt
@@ -27,5 +27,5 @@ data class BatchCostPriceEntity(
      * так как при распределении на 100 кг получается меньше 0.5 копеек на 1 грамм
      */
     @ColumnInfo(name = "cost_price_per_unit")
-    val costPricePerUnit: Int
+    val costPricePerUnit: Long
 )
