+++ database/src/desktopMain/kotlin/ru/pavlig43/database/data/product/Product.kt
@@ -30,7 +30,7 @@ data class Product(
     val comment: String = "",
 
     @ColumnInfo("price_for_sale")
-    val priceForSale: Int = 0,
+    val priceForSale: Long = 0,
 
     @PrimaryKey(autoGenerate = true)
     override val id: Int = 0,
