+++ database/src/desktopMain/kotlin/ru/pavlig43/database/data/safety/SafetyTableItem.kt
@@ -5,7 +5,7 @@ data class SafetyTableItem(
     val productId: Int,
     val productName: String,
     val vendorName: String,
-    val count: Int,
-    val reorderPoint: Int,
-    val orderQuantity: Int,
+    val count: Long,
+    val reorderPoint: Long,
+    val orderQuantity: Long,
 )
