+++ database/src/desktopMain/kotlin/ru/pavlig43/database/data/transact/buy/BuyBD.kt
@@ -46,7 +46,7 @@ data class BuyBDIn(
     val movementId: Int,
 
     @ColumnInfo("price")
-    val price: Int, // Цена в копейках
+    val price: Long, // Цена в копейках
 
     @ColumnInfo("comment")
     val comment: String,
@@ -77,7 +77,7 @@ data class BuyBDIn(
  */
 data class BuyBDOut(
     val transactionId: Int,
-    val count: Int,
+    val count: Long,
     val batchId: Int,
     val movementId: Int,
     val productId: Int,
@@ -86,7 +86,7 @@ data class BuyBDOut(
     val declarationName: String,
     val vendorName: String,
     val dateBorn: LocalDate,
-    val price: Int,
+    val price: Long,
     val comment: String,
     override val id: Int
 ) : CollectionObject
